requirements.txt
pandas>=1.3.0
matplotlib>=3.5.0
python-dotenv>=0.20.0
streamlit>=1.18.0
plotly>=5.10.0


# NOTE: The following libraries are used for feeder.
# TODO: When move to FAST API, remove these libraries.
psycopg2
sqlalchemy
praw
yfinance
gnews
torch
transformers
geopy
nltk